requirements/pip_upgrade.txt
Pillow # ==5.3.0
Pillow==6.0.0
Sphinx # ==1.8.2
Sphinx==2.0.1
Werkzeug # ==0.14.1
Werkzeug==0.15.2
argon2-cffi # ==18.3.0
argon2-cffi==19.1.0
coreapi # ==2.3.3
coreapi==2.3.3
coverage # ==4.5.2
coverage==4.5.3
django # ==2.1.4
Django==2.2.1
django-allauth # ==0.38.0
django-allauth==0.39.1
django-anymail[mailgun] # ==5.0
django-anymail==6.0
django-compressor # ==2.2
django-compressor==2.2
django-coverage-plugin # ==1.6.0
django-coverage-plugin==1.6.0
django-crispy-forms # ==1.7.2
django-crispy-forms==1.7.2
django-debug-toolbar # ==1.11
django-debug-toolbar==1.11
django-environ # ==0.4.5
django-environ==0.4.5
django-extensions # ==2.1.4
django-extensions==2.1.6
django-model-utils # ==3.1.2
django-model-utils==3.1.2
django-redis # ==4.10.0
django-redis==4.10.0
django-storages[boto3] # ==1.7.1
django-storages==1.7.1
djangorestframework # ==3.9.0
djangorestframework==3.9.3
factory-boy # ==2.11.1
factory-boy==2.11.1
flake8 # ==3.6.0
flake8==3.7.7
gunicorn # ==19.9.0
gunicorn==19.9.0
ipdb # ==0.11
ipdb==0.12
mypy # ==0.650
mypy==0.701
psycopg2 # ==2.7.6.1
psycopg2==2.8.2
psycopg2-binary # ==2.7.6.1
psycopg2-binary==2.8.2
pytest # ==4.0.1
pytest==4.4.1
pytest-django # ==3.4.4
pytest-django==3.4.8
pytest-sugar # ==0.9.2
pytest-sugar==0.9.2
python-slugify # ==1.2.6
python-slugify==3.0.2
pytz # ==2018.7
pytz==2019.1
rcssmin # ==1.0.6
rcssmin==1.0.6
redis # >=2.10.6,
redis==3.2.1
sendgrid # ==5.6.0
sendgrid==6.0.5
whitenoise # ==4.1.2
whitenoise==4.1.2

Package                       Version
----------------------------- ----------
Babel==2.6.0
Django==2.2.1
Faker==1.0.1
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==6.0.0
Pygments==2.3.0
Sphinx==2.0.1
Unidecode==1.0.23
Werkzeug==0.15.2
alabaster==0.7.12
appnope==0.1.0
argon2-cffi==19.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
boto3==1.9.142
botocore==1.12.142
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
decorator==4.3.0
defusedxml==0.5.0
django-allauth==0.39.1
django-anymail==6.0
django-appconf==1.0.2
django-compressor==2.2
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.6
django-model-utils==3.1.2
django-redis==4.10.0
django-storages==1.7.1
djangorestframework==3.9.3
docutils==0.14
entrypoints==0.3
factory-boy==2.11.1
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
ipdb==0.12
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
jmespath==0.9.4
mccabe==0.6.1
more-itertools==4.3.0
mypy==0.701
mypy-extensions==0.4.1
oauthlib==2.1.0
packaging==18.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pip==19.1
pluggy==0.9.0
prompt-toolkit==2.0.7
psycopg2==2.8.2
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.3.0
pytest==4.4.1
pytest-django==3.4.8
pytest-sugar==0.9.2
python-dateutil==2.7.5
python-http-client==3.1.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
rcssmin==1.0.6
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.0.0
rjsmin==1.0.12
s3transfer==0.2.0
sendgrid==6.0.5
setuptools==40.6.3
six==1.12.0
snowballstemmer==1.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
wheel==0.32.3
whitenoise==4.1.2
